Other factors move to Razor page with tests Enable runtime view changes for local development (launch settings)
