--- Frontend.Tests/ControllerTests/Projects/BenefitsControllerTests.cs
@@ -1,314 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Data;
-using Data.Models;
-using Data.Models.Projects;
-using Frontend.Controllers.Projects;
-using Frontend.Models;
-using Frontend.Models.Benefits;
-using Frontend.Tests.Helpers;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Routing;
-using Moq;
-using Xunit;
-
-namespace Frontend.Tests.ControllerTests.Projects
-{
-    public class BenefitsControllerTests
-    {
-        private readonly BenefitsController _subject;
-        private readonly Mock<IProjects> _projectsRepository;
-        private readonly Project _foundProject;
-        private const string Urn0001 = "0001";
-
-        public BenefitsControllerTests()
-        {
-            _foundProject = new Project
-            {
-                Urn = Urn0001,
-                TransferringAcademies = new List<TransferringAcademies>
-                {
-                    new TransferringAcademies()
-                    {
-                        OutgoingAcademyUrn = "Outgoing Academy Urn"
-                    }
-                }
-            };
-
-            _projectsRepository = new Mock<IProjects>();
-
-            _projectsRepository.Setup(r => r.GetByUrn(It.IsAny<string>()))
-                .ReturnsAsync(new RepositoryResult<Project> {Result = _foundProject});
-
-            _projectsRepository.Setup(r => r.Update(It.IsAny<Project>()))
-                .ReturnsAsync(new RepositoryResult<Project>());
-
-            _subject = new BenefitsController(_projectsRepository.Object);
-        }
-        
-        public class OtherFactorsTests : BenefitsControllerTests
-        {
-            public class GetTests : OtherFactorsTests
-            {
-                [Fact]
-                public async void GivenUrn_AssignsModelToTheView()
-                {
-                    var result = await _subject.OtherFactors(Urn0001);
-                    var viewModel = ControllerTestHelpers.AssertViewModelFromResult<OtherFactorsViewModel>(result);
-
-                    Assert.Equal(_foundProject.Urn, viewModel.Urn);
-                }
-
-                [Fact]
-                public async void GivenReturnToPreview_AssignsToTheModel()
-                {
-                    var result = await _subject.OtherFactors(Urn0001, true);
-                    var viewModel = ControllerTestHelpers.AssertViewModelFromResult<OtherFactorsViewModel>(result);
-
-                    Assert.True(viewModel.ReturnToPreview);
-                }
-
-                [Fact]
-                public async void GivenGetByUrnReturnsError_DisplayErrorPage()
-                {
-                    _projectsRepository.Setup(r => r.GetByUrn(It.IsAny<string>()))
-                        .ReturnsAsync(new RepositoryResult<Project>
-                        {
-                            Error = new RepositoryResultBase.RepositoryError
-                            {
-                                StatusCode = System.Net.HttpStatusCode.NotFound,
-                                ErrorMessage = "Project not found"
-                            }
-                        });
-
-                    var controller = new BenefitsController(_projectsRepository.Object);
-
-                    var response = await controller.OtherFactors("projectUrn");
-                    var viewResult = Assert.IsType<ViewResult>(response);
-                    var viewModel = ControllerTestHelpers.AssertViewModelFromResult<string>(response);
-
-                    Assert.Equal("ErrorPage", viewResult.ViewName);
-                    Assert.Equal("Project not found", viewModel);
-                }
-            }
-
-            public class PostTests : OtherFactorsTests
-            {
-                [Fact]
-                public async void GivenUrnAndNoOtherFactors_UpdateTheProject()
-                {
-                    var vm = new OtherFactorsViewModel
-                    {
-                        Urn = Urn0001,
-                        ReturnToPreview = false,
-                        OtherFactorsVm = new List<OtherFactorsItemViewModel>()
-                    };
-
-                    await _subject.OtherFactorsPost(vm);
-
-                    _projectsRepository.Verify(r =>
-                        r.Update(It.Is<Project>(project => !project.Benefits.OtherFactors.Any()))
-                    );
-                }
-
-                [Fact]
-                public async void GivenUrnAndAllOtherFactors_UpdateTheProject()
-                {
-                    Func<Project, bool> assertOtherFactorsEqual = project =>
-                    {
-                        var projectOtherFactors = project.Benefits.OtherFactors;
-                        var highProfile = projectOtherFactors[TransferBenefits.OtherFactor.HighProfile];
-                        var complexIssues =
-                            projectOtherFactors[TransferBenefits.OtherFactor.ComplexLandAndBuildingIssues];
-                        var finance = projectOtherFactors[TransferBenefits.OtherFactor.FinanceAndDebtConcerns];
-
-                        return highProfile == "High profile" &&
-                               complexIssues == "Complex issues" &&
-                               finance == "Finance concerns";
-                    };
-
-                    var otherFactors = new List<OtherFactorsItemViewModel>
-                    {
-                        new OtherFactorsItemViewModel
-                        {
-                            OtherFactor = TransferBenefits.OtherFactor.HighProfile,
-                            Checked = true,
-                            Description = "High profile"
-                        },
-                        new OtherFactorsItemViewModel
-                        {
-                            OtherFactor = TransferBenefits.OtherFactor.FinanceAndDebtConcerns,
-                            Checked = true,
-                            Description = "Finance concerns"
-                        },
-                        new OtherFactorsItemViewModel
-                        {
-                            OtherFactor = TransferBenefits.OtherFactor.ComplexLandAndBuildingIssues,
-                            Checked = true,
-                            Description = "Complex issues"
-                        },
-                    };
-                    var vm = new OtherFactorsViewModel
-                    {
-                        OtherFactorsVm = otherFactors,
-                        Urn = Urn0001,
-                        ReturnToPreview = false
-                    };
-
-                    await _subject.OtherFactorsPost(vm);
-
-                    _projectsRepository.Verify(r =>
-                        r.Update(It.Is<Project>(project => assertOtherFactorsEqual(project)))
-                    );
-                }
-
-                [Fact]
-                public async void GivenUrnAndOtherFactors_RedirectsToTheSummaryPage()
-                {
-                    var otherFactors =
-                        Enum.GetValues(typeof(TransferBenefits.OtherFactor))
-                            .Cast<TransferBenefits.OtherFactor>()
-                            .Where(otherFactor => otherFactor != TransferBenefits.OtherFactor.Empty)
-                            .Select(otherFactor => new OtherFactorsItemViewModel()
-                            {
-                                OtherFactor = otherFactor,
-                                Checked = true,
-                                Description = "test description"
-                            }).ToList();
-                    var vm = new OtherFactorsViewModel
-                    {
-                        OtherFactorsVm = otherFactors,
-                        Urn = Urn0001,
-                        ReturnToPreview = false
-                    };
-
-                    var response = await _subject.OtherFactorsPost(vm);
-
-                    ControllerTestHelpers.AssertResultRedirectsToPage(response, "/Projects/Benefits/Index",
-                        new RouteValueDictionary(new {urn = Urn0001}));
-                }
-
-                [Theory]
-                [InlineData("HighProfile")]
-                [InlineData("ComplexLandAndBuildingIssues")]
-                [InlineData("FinanceAndDebtConcerns")]
-                public async void GivenUrnAndOtherFactor_UpdatesTheProjectCorrectly(string otherFactorString)
-                {
-                    var otherFactors = new List<OtherFactorsItemViewModel>
-                    {
-                        new OtherFactorsItemViewModel
-                        {
-                            OtherFactor =
-                                (TransferBenefits.OtherFactor) Enum.Parse(typeof(TransferBenefits.OtherFactor),
-                                    otherFactorString),
-                            Checked = true,
-                            Description = "test Description"
-                        }
-                    };
-                    var vm = new OtherFactorsViewModel
-                    {
-                        OtherFactorsVm = otherFactors,
-                        Urn = Urn0001,
-                        ReturnToPreview = false
-                    };
-
-                    await _subject.OtherFactorsPost(vm);
-                    _projectsRepository.Verify(r => r.Update(It.Is<Project>(
-                        project => project.Benefits.OtherFactors.Keys.Count == 1 &&
-                                   project.Benefits.OtherFactors[otherFactors[0].OtherFactor] ==
-                                   "test Description"
-                    )));
-                }
-
-                [Fact]
-                public async void GivenGetByUrnReturnsError_DisplayErrorPage()
-                {
-                    _projectsRepository.Setup(r => r.GetByUrn(It.IsAny<string>()))
-                        .ReturnsAsync(new RepositoryResult<Project>
-                        {
-                            Error = new RepositoryResultBase.RepositoryError
-                            {
-                                StatusCode = System.Net.HttpStatusCode.NotFound,
-                                ErrorMessage = "Project not found"
-                            }
-                        });
-
-                    var controller = new BenefitsController(_projectsRepository.Object);
-
-                    var otherFactors = new List<OtherFactorsItemViewModel>();
-                    var vm = new OtherFactorsViewModel
-                    {
-                        OtherFactorsVm = otherFactors,
-                        Urn = "projectUrn",
-                        ReturnToPreview = false
-                    };
-
-                    var response = await controller.OtherFactorsPost(vm);
-
-                    var viewResult = Assert.IsType<ViewResult>(response);
-                    var viewModel = ControllerTestHelpers.AssertViewModelFromResult<string>(response);
-                    Assert.Equal("ErrorPage", viewResult.ViewName);
-                    Assert.Equal("Project not found", viewModel);
-                }
-
-                [Fact]
-                public async void GivenUpdateReturnsError_DisplayErrorPage()
-                {
-                    _projectsRepository.Setup(r => r.Update(It.IsAny<Project>()))
-                        .ReturnsAsync(new RepositoryResult<Project>
-                        {
-                            Error = new RepositoryResultBase.RepositoryError
-                            {
-                                StatusCode = System.Net.HttpStatusCode.NotFound,
-                                ErrorMessage = "Project not found"
-                            }
-                        });
-
-                    var controller = new BenefitsController(_projectsRepository.Object);
-
-                    var otherFactors = new List<OtherFactorsItemViewModel>();
-                    var vm = new OtherFactorsViewModel
-                    {
-                        OtherFactorsVm = otherFactors,
-                        Urn = "projectUrn",
-                        ReturnToPreview = false
-                    };
-
-                    var response = await controller.OtherFactorsPost(vm);
-
-                    var viewResult = Assert.IsType<ViewResult>(response);
-                    var viewModel = ControllerTestHelpers.AssertViewModelFromResult<string>(response);
-                    Assert.Equal("ErrorPage", viewResult.ViewName);
-                    Assert.Equal("Project not found", viewModel);
-                }
-
-                [Fact]
-                public async void GivenReturnToPreview_RedirectToThePreviewPage()
-                {
-                    var otherFactors = new List<OtherFactorsItemViewModel>
-                    {
-                        new OtherFactorsItemViewModel
-                        {
-                            OtherFactor =
-                                (TransferBenefits.OtherFactor) Enum.Parse(typeof(TransferBenefits.OtherFactor),
-                                    "HighProfile"),
-                            Checked = true,
-                            Description = "Meow"
-                        }
-                    };
-                    var vm = new OtherFactorsViewModel
-                    {
-                        OtherFactorsVm = otherFactors,
-                        Urn = "projectUrn",
-                        ReturnToPreview = true
-                    };
-
-                    var result = await _subject.OtherFactorsPost(vm);
-                    ControllerTestHelpers.AssertResultRedirectsToPage(result, Links.HeadteacherBoard.Preview.PageName,
-                        new RouteValueDictionary(new {id = "projectUrn"}));
-                }
-            }
-        }
-    }
-}
